requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 8.3.1
keyboard == 0.13.5
matplotlib == 3.4.3
numpy == 1.19.5
opencv_python == 4.5.3.56
pandas == 1.3.2
scikit_learn == 0.24.2
tensorflow == 2.6.0
